requirements.txt
ollama-ocr
nltk
PyPDF2
langchain
ollama
langchain-community
langchain-google-genai
python-dotenv